requirements.txt
streamlit>=1.32.0
azure-ai-formrecognizer==3.3.3
azure-core>=1.26.0
azure-common>=1.1.28
openai>=1.35.7
python-dotenv>=1.0.1
